backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-telegram-bot
google-generativeai
python-multipart
psycopg2-binary
async-lru
ultralyticsplus==0.0.28
ultralytics
opencv-python-headless
huggingface-hub
httpx
python-magic
# Local ML dependencies (Issue #76)
torch
transformers
Pillow
firebase-functions
firebase-admin
a2wsgi
